[shifts] Remove week index feature
In practice, we will want to sync to the calendar of Paris and New York.
We might have holiday weeks that we organize totally differently. Having
a fixed week index in the code only limits us, this is a process done so
seldomly (only 52 weeks per year) that it can be done manually once per
year (or every couple of months).